% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{abbreviate_name}
\alias{abbreviate_name}
\title{Abbreviate a vector of names}
\usage{
abbreviate_name(name, max_word_length = Inf, collapse = "")
}
\arguments{
\item{name}{A character vector of long character strings.}

\item{max_word_length}{Maximum word length before truncation.}

\item{collapse}{A string used to join the abbreviation letters.}
}
\value{
A character vector of abbreviated names.
}
\description{
Abbreviate a vector of names
}
\examples{
abbreviate_name(c("Large Cap Growth", "Small Value"), max_word_length = 3)
}
