% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotWts.R
\name{barplotWts}
\alias{barplotWts}
\title{A Barplot of a Set of Portfolio Weights}
\usage{
barplotWts(
  wts.efront,
  legend.text = NULL,
  col = NULL,
  ylab = NULL,
  xlab = c("MU", "VOL"),
  bar.ylim = NULL,
  ...
)
}
\arguments{
\item{wts.efront}{Matrix of weights along the efficient frontier}

\item{legend.text}{Vector of text for the legend}

\item{col}{Vector of colors for the bars}

\item{ylab}{A label for the y axis}

\item{xlab}{A label for the x axis}

\item{bar.ylim}{Limits of the y axis for barplot}

\item{...}{additional parameters from barplot}
}
\value{
A barplot of efficient frontier weights
}
\description{
Uses the R barplot function to make a barplot of efficient 
frontier weights. See the manual page for barplot()
}
\examples{
args(barplotWts)
}
