% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BXMdata}
\alias{BXMdata}
\title{CBOE S&P 500 BuyWrite Index}
\format{
A data.frame object
}
\source{
TO BE ADDED.
}
\usage{
data(BXMdata)
}
\description{
Monthly data for the S&P500 index, the S&P 500 BuyWrite Index, and the risk-free rate
}
\examples{
library(PCRA)
names(BXMdata)
}
\keyword{datasets}
