% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConferenceBoardETI.R
\docType{data}
\name{ConferenceBoardETI}
\alias{ConferenceBoardETI}
\title{ConferenceBoardETI}
\format{
A data frame with a time series of month-end observations of the
Conference Board Employment Trends Index (ETI) from November 1973 to 
December 2023.
									
\itemize{
 \item \strong{Date:} type `Date`. Last Day of Month. The time series extends 
 from November 1973 to December 2023.
 \item \strong{ETI:} type `num`. Level of the Conference Board Employment 
 Trends Index on the last day of the month.
}
}
\source{
The Conference Board. The Conference Board ETI Data is provided by the
Conference Board for educational purposes only. The Conference Board makes no 
representation or warranty of any kind, express or implied, including regarding 
the accuracy, adequacy, validity, reliability, availability or completeness of 
The Conference Board ETI Data. All rights reserved. Redistribution of the data 
is not permitted, and use use of the data in derivative works is not permitted 
without the written permission of the Conference Board.
}
\usage{
data(ConferenceBoardETI)
}
\description{
Monthly time series of the Conference Board Employment Trends
Index from November 1973 to December 2023.
}
\details{
This data set provides a monthly time series of the Conference Board 
Employment Trends Index (ETI) from November 1973 to December 2023. Details of
the index's construction as well as its most recent value can be found at 
https://www.conference-board.org/data/eti.cfm. Technical details related to the 
construction of the index are available at 
https://www.conference-board.org/pdf_free/press/Employment%20Trends%20Index%20Technical%20Notes65.pdf
while details of its revision in November 2020 can be found at
https://www.conference-board.org/pdf_free/press/Employment%20Trends%20Index%20Methodology%20Revision%20November%209%20202023.pdf
The change was made following the COVID-19 driven recession in 2020. 
The Conference Board ETI combines 8 measures of employment to create a composite
index of employment trends.
1. Percentage of respondents who say they find “Jobs Hard to Get” 
(Source: The Conference Board Consumer Confidence Survey)
2. Initial Claims for Unemployment Insurance, State Programs 
(Source: U.S. Department of Labor).
3. Percentage of Firms With Positions Not Able to Fill Right Now 
(Source: © National Federation of Independent Business Research Foundation)
4. Number of Employees Hired by the Temporary-Help Industry 
(Source: U.S. Bureau of Labor Statistics)
5. Part-time Workers for Economic Reasons (Bureau of Labor Statistics)
6. Job Openings (Bureau of Labor Statistics, through Job Openings and 
Labor Turnover Survey):
7. Industrial Production (Federal Reserve Board) 
8. Real Manufacturing and Trade Sales (U.S. Bureau of Economic Analysis).
}
\examples{
 
data(ConferenceBoardETI)
names(ConferenceBoardETI)
head(ConferenceBoardETI, 5)
tail(ConferenceBoardETI, 5)
}
\references{
Chapter 13 (Expected Returns) of Martin, Philips, Scherer, 
Stoyanov and Li, Portfolio Construction and Risk Analysis, Springer, 2024.
}
\keyword{datasets}
