% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huberRhoPsiWtsEff.R
\name{constFromEffHuber}
\alias{constFromEffHuber}
\title{Constant from Efficiency Huber Psi}
\usage{
constFromEffHuber(eff, interval = c(1e-06, 3))
}
\arguments{
\item{eff}{}

\item{interval}{}
}
\description{
Constant from Efficiency Huber Psi
}
\examples{
args(constFromEffHuber)
}
\seealso{
\code{\link{}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{huber}
