% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datFF3W}
\alias{datFF3W}
\title{Fama-French 3-Factor Model Weekly Time Series}
\format{
Multivariate time series xts object
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
data(datFF3W)
}
\description{
Weekly values of the 3 factors MKT, SMB and HML
}
\examples{
library(PCRA)
library(zoo)
data(datFF3W)
head(datFF3W)
range(index(datFF3W))
}
\keyword{datasets}
