% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datFF4W}
\alias{datFF4W}
\title{Fama-French 4-Factor Model Weekly Time Series}
\format{
Multivariate time series xts object
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
data(datFF4W)
}
\description{
Weekly values of the 4 factors MKT, SMB, HML and MOM
}
\examples{
library(PCRA)
library(zoo)
data(datFF4W)
head(datFF4W)
range(index(datFF4W))
}
\keyword{datasets}
