% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divHHI.R
\name{divHHI}
\alias{divHHI}
\title{divHHI calculates a portfolio diversification index DIV}
\usage{
divHHI(weights)
}
\arguments{
\item{weights}{}
}
\value{
Time series of portfolio diversification values
}
\description{
DIV is one minus the Herfindahl-Hirschman Index (HHI)index, which is defined
to be the sum of the squared portfolio weights.
}
\details{
The maximum HHI of a long-only portfolio is 1, which occurs when all of the
portfolio's investment is in a single asset, and correspondingly HHI = 0.
}
\examples{
args(divHHI)
}
