% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfPointMassFig.R
\name{edfPointMassFig}
\alias{edfPointMassFig}
\title{edfPointMassFig}
\usage{
edfPointMassFig(r, x1 = -0.02, x2 = 0.02)
}
\arguments{
\item{r}{Numeric value for location of point mass}

\item{x1}{Lower limit of plotting range}

\item{x2}{Upper limit of plotting range}
}
\value{
A figure of the point mass distribution function
}
\description{
This is code for plotting a point mass distribution function
}
\examples{
edfPointMassFig(.005)
}
