% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsesPlotNewKjell.covfm.R
\name{ellipsesPlotNewKjell.covfm}
\alias{ellipsesPlotNewKjell.covfm}
\title{Ellipses Plot for fit.models}
\usage{
ellipsesPlotNewKjell.covfm(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
A Figure
}
\description{
Ellipses Plot for fit.models
}
\examples{
args(ellipsesPlotNewKjell.covfm)
}
