% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewmaMeanVol.R
\name{ewmaMeanVol}
\alias{ewmaMeanVol}
\title{EWMA Mean and Volatility}
\usage{
ewmaMeanVol(
  x,
  nstart = 10,
  robMean = T,
  robVol = T,
  cc = 2.5,
  lambdaMean = 0.9,
  lambdaVol = 0.9
)
}
\arguments{
\item{x}{An xts returns object}

\item{nstart}{Integer number of returns for initial ewma estimates}

\item{robMean}{Logical variable, if TRUE compute robust ewmaMean,
if FALSE compute classic classic ewmaMean. Default is TRUE}

\item{robVol}{Logical variable, if TRUE compute robust ewmaVol,
if FALSE compute classic ewmaVol. Default is TRUE.}

\item{cc}{Numeric value, robustness tuning constant. Default is 2.5}

\item{lambdaMean}{Numeric value, decay rate constant for ewmaMean}

\item{lambdaVol}{Numeric value, decay rate constant for ewmaVol}
}
\value{
A bivarate xts object containing the ewmaMean and ewmaVol
}
\description{
Computes EWMA joint mean and vol, with options for
robust versus classic EWMA estimates.
}
\examples{
args(ewmaMeanVol)
}
