% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewmaVol.R
\name{ewmaVol}
\alias{ewmaVol}
\title{Exponential Weighted Moving Average (EWMA) Volitility}
\usage{
ewmaVol(x, nstart = 10, robVol = TRUE, cc = 2.5, lambda = 0.9)
}
\arguments{
\item{x}{returns of the portfolio}

\item{nstart}{startup values, default is nstart = 10}

\item{robVol}{if the robust vol is used, default is TRUE}

\item{cc}{robustness tuning constant, default 2.5}

\item{lambda}{EWMA decay parameter, default 0.9}
}
\value{

}
\description{
Exponential Weighted Moving Average (EWMA) Volitility
}
