% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FRBinterestRates}
\alias{FRBinterestRates}
\title{Federal Reserve Board Interest Rates}
\format{
A time series zoo object
}
\source{
Federal Reserve Board
}
\usage{
data(FRBinterestRates)
}
\description{
Federal Reserve Board monthly interest rates of 90 day Bill from 1934
to 2014.
}
\examples{
library(PCRA)
library(zoo)
data(FRBinterestRates)
class(FRBinterestRates)
range(index(FRBinterestRates))
}
\keyword{datasets}
