% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gfunds5}
\alias{gfunds5}
\title{gfunds5}
\format{
Multivariate xts object
}
\source{
Unknown
}
\usage{
data(gfunds5)
}
\description{
Monthly returns of 5 German investment funds November 1989
to July 2001: EM (emerging markets), PE (private equity), HY (high yield),
ALT (alternatives), and BND (fixed income)
}
\examples{
library(PCRA)
data(gfunds5)
class(gfunds5)
names(gfunds5)
range(index(gfunds5))
}
\keyword{datasets}
