% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HFstrategies}
\alias{HFstrategies}
\title{Hedge Fund Strategies Returns}
\format{
A multivariate xts object
}
\source{
Unknown
}
\usage{
data(HFstrategies)
}
\description{
Monthly returns of 9 hedge fund strategies from 1994 to 2004
}
\examples{
library(PCRA)
library(zoo)
data(HFstrategies)
names(HFstrategies)
dim(HFstrategies)
range(index(HFstrategies))
}
\keyword{datasets}
