% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invensysEPS}
\alias{invensysEPS}
\title{Earnings per Share of Invensys}
\format{
A numeric vector
}
\source{
Corporate Finance Department of Dupont
}
\usage{
data(invensysEPS)
}
\description{
Yearly earnings-per-share of company Invensys for 17 years.
The company's name was invensys prior to 2004.
}
\examples{
library(PCRA)
data(invensysEPS)
invensysEPS
}
\keyword{datasets}
