% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRest.R
\name{KRest}
\alias{KRest}
\title{Kurtosis Estimator}
\usage{
KRest(x, excess = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{excess}{A logical variable with default TRUE, which results in the
computation of excess kurtosis, and FALSE results ordinary kurtosis.}
}
\value{
The estimate of excess kurtosis or ordinary kurtosis
}
\description{
Sample estimate of excess kurtosis, with option for ordinary
kurtosis. This function will eventually have a robust estimator option.
}
\examples{
args(KRest)
}
