% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levgLongShort.R
\name{levgLongShort}
\alias{levgLongShort}
\title{Long Short Portfolio Leverage}
\usage{
levgLongShort(w)
}
\arguments{
\item{w}{a multivariate xts portfolio weights object}
}
\value{
levg an xts time series of portfolio leverages
}
\description{
This function computes a time series of portfolio leverages, defined  as
the sum of the absolute portfolio weights divided by the sum of the long
position weights
}
\examples{


}
\author{
Doug Martin, \email{martinrd3d@gmail.com}
}
