% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MarketData}
\alias{MarketData}
\title{Global Baskets of Equity and Bonds}
\format{
An xts object
}
\source{
TO BE ADDED.
}
\usage{
data(MarketData)
}
\description{
Daily price data for a global equity basket, a global bond basket, and cash
}
\examples{
library(PCRA)
range(range(MarketData$global.equity.basket))
}
\keyword{datasets}
