% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathEfrontCashRisky.R
\name{mathEfrontCashRisky}
\alias{mathEfrontCashRisky}
\title{mathEfrontCashRisky}
\usage{
mathEfrontCashRisky(
  returns,
  npoints = 10,
  rf = 0.005,
  plot.efront = TRUE,
  stock.names = TRUE,
  values = FALSE,
  scalex = 1.1,
  scaley = 1.1,
  cexPoints = 0.8,
  cexText = 0.8
)
}
\arguments{
\item{returns}{Risky asset returns in the form of an xts object}

\item{npoints}{Number of points at which to compute efront values and
weights}

\item{rf}{A risk-free rate}

\item{plot.efront}{Logical variable which if TRUE results in a plot of
the linear efficient frontier and the stock positions}

\item{stock.names}{Logical variable which if TRUE results in stock names 
being displayed}

\item{scalex}{A scale parameter for scaling the horizontal axis}

\item{scaley}{A scale parameter for scaling the vertical axis}

\item{digits}{Number of digits for printing efront weights}
}
\value{
A plot of the linear efficient frontier or the weights along the
efficient frontier.  Optionally output the mean and vol values
of the linear efficient frontier or its weights, or both.
}
\description{
This function computes and plots a linear efficient frontier that is
a mix of a risk-free asset ("cash") and risky stocks (or other assets).
It optionally returns the weights along the linear efficient frontier.
}
\examples{
args(mathEfrontCashRisky)
}
