% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathGmv.R
\name{mathGmv}
\alias{mathGmv}
\title{Global Minimum Variance Portfolio (GMV)}
\usage{
mathGmv(returns, digits = NULL)
}
\arguments{
\item{returns}{Matrix of asset returns}

\item{digits}{Integer value of number of significant digits, default NULL}
}
\value{
List of GMV portfolio weights, mean return and volatility
}
\description{
Computes the weights of a GMV portfolio, and its mean return
and volatility
}
\examples{
args(mathGmv)
}
