% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathGmvMuCov.R
\name{mathGmvMuCov}
\alias{mathGmvMuCov}
\title{Global Minimum Variance Portfolios From Mu and Cov}
\usage{
mathGmvMuCov(muRet, volRet, corrRet, digits = 3)
}
\arguments{
\item{muRet}{Mean vector}

\item{volRet}{Volatility vector}

\item{corrRet}{matrix of correlations}

\item{digits}{Integer value number of decimal places, default 3}
}
\value{
a list contains weights, mean return and volatility of a GMV portfolio
}
\description{
Compute the weights, mean return and volatility of a GMV portfolio based on
user specified mean vector and covariance matrix
}
\examples{
args(mathGmvMuCov)
}
