% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathTport.R
\name{mathTport}
\alias{mathTport}
\title{Tangency Portfolio Weights}
\usage{
mathTport(returns, rf = 0.005, digits = NULL)
}
\arguments{
\item{returns}{A vector or xts object}

\item{rf}{The risk-free rate, default 0.005}

\item{digits}{Number of significant digits default NULL}
}
\value{
Tangency portfoliow weights, mean and volatility
}
\description{
Computes the portfolio weights of the tangency portfolio, and
its mean return and volatility.  The tangency portfolio is defined by the
line connecting the zero volatility risk-free rate to its tangency point on
the efficient frontier.
}
\examples{
args(mathTport)
}
