% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathWtsEfrontRisky.R
\name{mathWtsEfrontRisky}
\alias{mathWtsEfrontRisky}
\title{Efficient Frontier Portfolio Weights}
\usage{
mathWtsEfrontRisky(returns, mu.efront, digits = NULL)
}
\arguments{
\item{returns}{A multivariate xts object of n asset returns}

\item{mu.efront}{A vector of specified efficient frontier mean returns}

\item{digits}{Integer number of significant digits for output}
}
\value{
A matrix with first row containing the mean (MU) along the efficient
frontier, the second row containing the standard deviation, and the 
following n rows contain the n weight vectors along the efficient fronier
}
\description{
Computes weights vector for a set of points along the efficient
frontier that are defined by their mean return values
}
\examples{
args(mathWtsEfrontRisky)

}
