% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathWtsEfrontRiskyMuCov.R
\name{mathWtsEfrontRiskyMuCov}
\alias{mathWtsEfrontRiskyMuCov}
\title{Portfolio Weights on an Efficient Frontier}
\usage{
mathWtsEfrontRiskyMuCov(muRet, volRet, corrRet, mu.efront, digits = NULL)
}
\arguments{
\item{muRet}{Vector of asset mean returns}

\item{volRet}{Vector of asset volatilies}

\item{corrRet}{Asset correlation matrix}

\item{mu.efront}{A vector of specified efficient frontier mean returns}

\item{digits}{If null no rounding of weights, else number of significant
digits for rounding}
}
\value{
A matrix whose first row contains the volatilities along the efficient
at specified mean returns contained in the second row, and the remaining
n rows contain contain the n weight vectors at each of the efficient
frontier mean return values
}
\description{
Same as function \code{"mathWtsEfrontRisky"} except that instead a user specified
time series of portfolio return, it is based on user specified returns 
mean vector and covariance matrix
}
\examples{
args(mathWtsEfrontRiskyMuCov)
}
