% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanRetReport.R
\name{meanRetReport}
\alias{meanRetReport}
\title{Four types of returns means for reporting}
\usage{
meanRetReport(x, robust = FALSE)
}
\arguments{
\item{x}{An xts or numeric vector of return}

\item{robust}{A logical value controlling whether a classical or robust sample
mean and standard deviation is computed. Default is FALSE}
}
\value{
Arithmetic, logarithmic, geometric and approximate geometric means
}
\description{
Computes arithmetic mean, logarithmic mean, geometric mean and an approximate
geometric mean.
}
\examples{
args(meanRetReport)
}
