% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanReturns4Types.R
\name{meanReturns4Types}
\alias{meanReturns4Types}
\title{Four Types of Mean Returns}
\usage{
meanReturns4Types(ret, robust = FALSE, eff = 0.95)
}
\arguments{
\item{robust}{A logical value controlling whether a classical or robust sample
mean and standard deviation is computed. Default is FALSE}

\item{x}{An xts or numeric vector of returns}
}
\value{
fourMeans
}
\description{
Computation of arithmetic mean, logarithmic mean, geometric,
mean and an approximate
geometric mean.
}
\examples{
args(meanReturns4Types)
}
