% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMarketData.R
\name{mergeMarketData}
\alias{mergeMarketData}
\title{Merge stocksCRSP and factorsSPGMI}
\usage{
mergeMarketData(
  stocks = stocksCRSP,
  factors = factorsSPGMI,
  dateSet = c("1993-01-31", "2015-12-31"),
  stockItems = c("Date", "TickerLast", "CapGroup", "Sector", "Return", "Ret13WkBill",
    "mktIndexCRSP"),
  factorItems = c("BP", "LogMktCap", "SEV"),
  tickerList = NULL
)
}
\arguments{
\item{stocks}{A data.table of stock returns and related data}

\item{factors}{A data.table of factors and related data}

\item{dateSet}{A character vector providing a start data and an end 
date, having the same form as c("2006-01-31", "2010-12-31")}

\item{stockItems}{A character vector that is a subset of the names
of stocks data.table}

\item{factorItems}{A character vector that is a subset of the names
of factors data.table}

\item{tickerList}{A character vector that contains a subset of tickerLast 
items}
}
\value{
A merged data.table consisting of selected stocks and factors
}
\description{
Select and merge a set of CRSP stocks and SPGMI factors
over a range of dates

Select and merge a set of CRSP stocks and SPGMI factors
over a range of dates

Select and merge a set of CRSP stocks and SPGMI factors and
SPGMI factors over a range of dates
}
\details{
The date range, stock data items, and factor items have default values but
can be changed by the user
}
\examples{
data(stocksCRSP)
data(factorsSPGMI)

stocks_factors <- selectCRSPandSPGMI(stocks = stocksCRSP, factors = factorsSPGMI,
                                    dateSet = c("2006-01-31", "2010-12-31"), 
                                    stockItems = c("Date", "TickerLast", 
                                                   "CapGroup", "Sector", 
                                                   "Return", "Ret13WkBill",
                                                   "mktIndexCRSP"),
                                    factorItems = c("BP", "LogMktCap", "SEV"),
                                    tickerList = c("ABT","ABM","MMM","NKE")
                                    )

str(stocks_factors)
}
