% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minVarCashRisky.R
\name{minVarCashRisky}
\alias{minVarCashRisky}
\title{Minimum Variance Portfolio}
\usage{
minVarCashRisky(returns, mu0, LO = FALSE, bnd = 1000, bndRF = 100)
}
\arguments{
\item{returns}{xts multivariate returns object that contains
the returns of the risk-free T-Bill in the last column}

\item{mu0}{Minimum variance portfolio mean return}

\item{LO}{Logical variable with default FALSE}

\item{bnd}{Bound on asset weights}

\item{bndRF}{a bound on the risk-free T-Bill weight}
}
\value{
A list of the minimum variance portfolio numeric weights, 
mean value, volatility, and Sharpe Ratio.
}
\description{
Given a time series of asset returns and risk-free T-Bill
returns, a target mean return, and a specification of whether the asset
weights are constrained to be long-only, this function compputes the 
weights of a fully-invested minimum variance portfolio
}
\examples{
args(minVarCashRisky)
}
