% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLSandHuberRobustSFM.R
\name{plotLSandHuberRobustSFM}
\alias{plotLSandHuberRobustSFM}
\title{Plot LS and Huber SFM Fits}
\usage{
plotLSandHuberRobustSFM(
  x,
  mainText = NULL,
  ylimits = NULL,
  legendPos = "topleft",
  goodOutlier = FALSE,
  makePct = FALSE
)
}
\arguments{
\item{x}{xts time series vector}

\item{mainText}{Character variable with NULL default}

\item{ylimits}{Numeric vector of vertical axis limits with NULL default}

\item{legendPos}{Character variable with default "topleft"}

\item{goodOutlier}{Logical variable with default FALSE}

\item{makePct}{Logical variable with default FALSE}
}
\value{
A plot of the LS and robust Huber SFM fits
}
\description{
Computes LS and Huber robust single factor model fits with 
standard errors and plots the results
}
\examples{
args(plotLSandHuberRobustSFM)
}
