% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLSandRobustSFM.R
\name{plotLSandRobustSFM}
\alias{plotLSandRobustSFM}
\title{Plot LS and Robust SFM Fits}
\usage{
plotLSandRobustSFM(
  x,
  family = "mopt",
  efficiency = 0.95,
  mainText = NULL,
  ylimits = NULL,
  legendPos = "topleft",
  goodOutlier = F,
  makePct = FALSE
)
}
\arguments{
\item{x}{}

\item{family}{}

\item{efficiency}{}

\item{mainText}{}

\item{ylimits}{}

\item{legendPos}{}

\item{goodOutlier}{}

\item{makePct}{}
}
\value{

}
\description{
Plot LS and Robust SFM Fits
}
