% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerUtilityPlots.R
\name{powerUtilityPlots}
\alias{powerUtilityPlots}
\title{Power Utility Function Plots}
\usage{
powerUtilityPlots()
}
\value{
A figue of the above three utility functions overlaid
}
\description{
Uses the power utility function formulas to plot power utility 
functions for gamma = 0.5, gamma = -0.5 and the log utility function
limiting case of a power utility function as gamma -> 0.
 NOTE: Need to add  LaTex form of the power utility function formula
}
\examples{
args(powerUtilityPlots)
}
