% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huberRhoPsiWtsEff.R
\name{psiHardRej}
\alias{psiHardRej}
\title{Psi Hard Rejection}
\usage{
psiHardRej(x, cc = 2.5)
}
\arguments{
\item{x}{a numeric vector}

\item{cc}{a numeric constant}
}
\description{
Psi Hard Rejection
}
\examples{
args(psiHardRej)
}
\seealso{
\code{\link{}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{huber}
