% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiHuber.R
\name{psiHuber}
\alias{psiHuber}
\title{Huber psi function}
\usage{
psiHuber(x, cc = 2)
}
\arguments{
\item{x}{Numeric argument of psi function}

\item{cc}{numeric robustness tuning constant}
}
\value{
Numeric value of psi(x)
}
\description{
The Huber psi function psi(x) is the derivative
of the Huber rho function, with a tuning constant cc that 
controls the trade-off between robustness toward outliers, 
and normal distribution estimator efficiency.
}
\details{
The choice cc = 1.345 results in a 95% normal 
distribution efficiency for the Huber location M-estimator.
The default value cc = 2.0 is better for EWMA robust filters.
}
\examples{
psiHuber(0.5)
}
