% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqnormDatWindat.R
\name{qqnormDatWindat}
\alias{qqnormDatWindat}
\title{qqnormDatWindat}
\usage{
qqnormDatWindat(
  dat,
  windat,
  fraction = 0.01,
  ylim = NULL,
  main = main,
  facName = NULL
)
}
\arguments{
\item{dat}{Numeric data vector}

\item{windat}{Numeric Winsorized data set}

\item{fraction}{Fraction of data that is Winsorized}

\item{ylim}{Numeric data with two values that control vertical plot range}

\item{main}{Character main title of plot}

\item{facName}{Character data for y axis label}
}
\value{
A normal QQPlot of data with overlaid Winsorized data
}
\description{
Normal QQPlot of data and Winsorized data
}
\details{
The result plot displays a normal QQPlot of the orginal data
as solid points, along with the horizontal display of the Winsorized
data as small circles.
}
\examples{
args(qqnormDatWindat)
}
