% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadraticUtilityPlot.R
\name{quadraticUtilityPlot}
\alias{quadraticUtilityPlot}
\title{Quadratic Utility Function Plot}
\usage{
quadraticUtilityPlot()
}
\value{
A plot of the quadratic utility
}
\description{
Plots the quadratic utility function U(v) = v - v^2.
NOTE:  Add LaTeX formula
}
\examples{
args(quadraticUtilityPlot)
}
