% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retDD}
\alias{retDD}
\title{CRSP Returns of Stock with Ticker DD}
\format{
Multivariate time series xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate
of the University of Chicago Booth School of Business.
}
\usage{
data(retDD)
}
\description{
Weekly returns (RET) of stock with ticker DD for 1986
and 1987, along with market returns (MKT) and risk-free rate (RF).
}
\examples{
library(PCRA)
library(zoo)
data(retDD)
head(retDD)
range(index(retDD))
}
\keyword{datasets}
