% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retEDS}
\alias{retEDS}
\title{CRSP Returns of Stock with Ticker EDS}
\format{
Multivariate time series xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business.
}
\usage{
data(retEDS)
}
\description{
Weekly returns (RET) of stock with ticker EDS for 2002
and 2003, along with market returns (MKT) and risk-free rate (RF).
}
\examples{
library(PCRA)
library(zoo)
data(retEDS)
head(retEDS)
range(index(retEDS))
}
\keyword{datasets}
