% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retFNB}
\alias{retFNB}
\title{CRSP Returns of Stock with Ticker FNB}
\format{
Univariate time series xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business.
}
\usage{
data(retFNB)
}
\description{
Weekly returns (RET) of stock with ticker FNB for 2008
}
\examples{
library(PCRA)
library(zoo)
data(retFNB)
head(retFNB)
range(index(retFNB))
}
\keyword{datasets}
