% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retKBH}
\alias{retKBH}
\title{Stock with Ticker KBH}
\format{
Multivariate time series xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business.
}
\usage{
data(retKBH)
}
\description{
Weekly returns (RET) of stock with ticker KBH for 2007
and 2008, along with market returns (MKT) and risk-free rate (RF).
}
\examples{
library(PCRA)
library(zoo)
data(retKBH)
head(retKBH)
range(index(retKBH))
}
\keyword{datasets}
