% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retMER}
\alias{retMER}
\title{Stock with Ticker MER}
\format{
Multivariate time series xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business.
}
\usage{
data(retMER)
}
\description{
Weekly returns (RET) of stock with ticker MER for 2002
and 2003, along with market returns (MKT) and risk-free rate (RF).
}
\examples{
library(PCRA)
library(zoo)
data(retMER)
head(retMER)
range(index(retMER))
}
\keyword{datasets}
