% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retOFG}
\alias{retOFG}
\title{CRSP Returns of Stock with Ticker OFG}
\format{
Multivariate time series xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business.
}
\usage{
data(retOFG)
}
\description{
Weekly returns (RET) of stock with ticker OFG for 2007
and 2008, along with market returns (MKT) and risk-free rate (RF).
}
\examples{
library(PCRA)
library(zoo)
data(retOFG)
head(retOFG)
range(index(retOFG))
}
\keyword{datasets}
