% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retPSC}
\alias{retPSC}
\title{CRSP Returns of Stock with Ticker PSC}
\format{
Multivariate time series xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business.
}
\usage{
data(retPSC)
}
\description{
Weekly returns (RET) of stock with ticker PSC for 1987
and 1088, along with market returns (MKT) and risk-free rate (RF).
}
\examples{
library(PCRA)
library(zoo)
data(retPSC)
head(retPSC)
range(index(retPSC))
}
\keyword{datasets}
