% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retVHI}
\alias{retVHI}
\title{Stock with Ticker VHI}
\format{
Multivariate time series xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business.
}
\usage{
data(retVHI)
}
\description{
Weekly returns (RET) of stock with ticker VHI for 1990
and 1991, along with market returns (MKT) and risk-free rate (RF).
}
\examples{
library(PCRA)
data(retVHI)
head(retVHI)
range(index(retVHI))
}
\keyword{datasets}
