% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retWTS}
\alias{retWTS}
\title{Stock with Ticker WTS}
\format{
Multivariate time series xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business.
}
\usage{
data(retWTS)
}
\description{
Weekly returns (RET) of stock with ticker WTS for 2009
and 2010, along with market returns (MKT) and risk-free rate (RF).
}
\examples{
library(PCRA)
library(zoo)
data(retWTS)
head(retWTS)
range(index(retWTS))
}
\keyword{datasets}
