% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huberRhoPsiWtsEff.R
\name{rhoHuber}
\alias{rhoHuber}
\title{Huber Rho}
\usage{
rhoHuber(x, cc = 1.345)
}
\arguments{
\item{x}{a numeric vector}

\item{cc}{a numeric constant}
}
\description{
Huber Rho
}
\examples{
args(rhoHuber)
}
\seealso{
\code{\link{}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{huber}
