% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCRSPandSPGMI.R
\name{selectCRSPandSPGMI}
\alias{selectCRSPandSPGMI}
\title{Merge and select subsets from stocksCRSP and factorsSPGMI data sets}
\usage{
selectCRSPandSPGMI(
  periodicity = "monthly",
  dateRange = c("1993-01-31", "2015-12-31"),
  stockItems = c("Date", "TickerLast", "CapGroupLast", "Sector", "Return", "Ret13WkBill",
    "MktIndexCRSP"),
  factorItems = c("BP", "LogMktCap", "SEV"),
  sectorList = NULL,
  capGroupList = NULL,
  tickerList = NULL
)
}
\arguments{
\item{periodicity}{Character "monthly","weekly","daily".}

\item{dateRange}{A character vector providing a start data and an end 
date, having the same form as c("2006-01-31", "2010-12-31").}

\item{stockItems}{A character vector that is a subset of the names
of columns in the stocksCRSP data.table}

\item{factorItems}{A character vector that is a subset of the names
of columns in the factorsSPGMI data.table}

\item{sectorList}{A character vector that is a subset of the sectors in the 
stocksCRSP dataset. Set to NULL to return all sectors.}

\item{capGroupList}{A character vector that is a subset of the CapGroupLast
("MicroCap","SmallCap","MidCap","LargeCap") variable. Set to NULL to return
all CapGroups.}

\item{tickerList}{A character vector that contains a subset of tickerLast 
items. Set to NULL to return all tickers.}
}
\value{
A data.table consisting of selected stocks and factor/stock data.
}
\description{
Select data from stocksCRSP (daily, weekly, and monthly) and 
merge with factorsSPGMI while also optionally selecting subsets by rows 
(selecting on date ranges, tickers, capitalization groups, and sectors or by 
columns (selecting a subset of the factor data for each ticker).
}
\details{
Users select a periodicity for the data (stocksCRSP is available in daily, 
weekly, and monthly variants) and this function will resample the monthly
factorsSPGMI data to the selected periodicity. 
Users may select all columns from both data sets, a specified set of columns,
or by setting either stockItems or factorItems to "NULL", may select only 
items from the other data set (that is, if only the stocksCRSP data is 
desired, set factorITEMS to NULL). 
Smaller sub-samples of the data (fewer rows) can be returned by
selecting a specific date range, sectors, CapGroups (MicroCap, SmallCap, 
etc.) of interest, or by specifying a list of tickers for which data can be
returned. Care should be taken to ensure consistency in combined 
restrictions when specifying CapGroup, Sector, and TickerLast values.
}
\examples{
data(stocksCRSP)
data(factorsSPGMI)

stocks_factors <- selectCRSPandSPGMI(periodicity = "monthly",
                                    dateRange = c("2006-01-31", "2010-12-31"), 
                                    stockItems = c("Date", "TickerLast", 
                                    "CapGroupLast", "Sector", "Return",
                                    "Ret13WkBill", "MktIndexCRSP"),
                                    factorItems = c("BP", "LogMktCap", "SEV"),
                                    sectorList = NULL,
                                    capGroupList = NULL,
                                    tickerList = NULL)

names(stocks_factors)
str(stocks_factors)
}
