% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifStars.R
\name{signifStars}
\alias{signifStars}
\title{signifStars}
\usage{
signifStars(pv)
}
\arguments{
\item{pv}{}
}
\value{
p-values with none, one, two, or three star symbols added
}
\description{
signifStars
}
\examples{
args(signifStars)
}
