% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKest.R
\name{SKest}
\alias{SKest}
\title{SKest}
\usage{
SKest(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Value of sample skewness
}
\description{
Sample estimate of skewness
This function will eventually have a robust estimation option
}
\examples{
args(SKest)
}
