% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocksCRSPdailyweekly.R
\docType{data}
\name{stocksCRSPdaily}
\alias{stocksCRSPdaily}
\title{CRSP daily stocks data for 294 stocks}
\format{
A data.table object with 82000 observations on 15 variables:
\itemize{
 \item \strong{Date:} type `Date`.
 \item \strong{TickerLast:} type `chr`. The ticker as of December 2015
 \item \strong{Ticker:} type `chr`. Monthly ticker
 period
 \item \strong{Company:} type `chr`. The name of company with TickerLast
 \item \strong{CapGroupLast:} type `chr`. Company market capitalization group
 as of December 2015, one of: MicroCap, SmallCap, MidCap or LargeCap
 \item \strong{CapGroup:} type `chr`. Monthly market capitalization group
 \item \strong{GICS:} type `chr`. 6 digit S&P GICS code
 \item \strong{Sector:} type `chr`. One of 10 sectors specified by the first
 two digits of the GICS code
 \item \strong{Return:} type `num`. Arithmetic stock return from one period
 to the next in decimal form
 \item \strong{RetExDiv:} type `num`.
 \item \strong{Price:} type `num`. Stock price at each time period in decimal
 form
 \item \strong{PrcSplitAdj:} type `num`.
 \item \strong{Ret4WkBill:} type `num`. Return of 4 week Treasury bill
 \item \strong{Ret13WkBill:} type `num`. Return of 13 week Treasury bill
 \item \strong{Ret1YrBill:} type `num`. Return of 1 year Treasury bill
 \item \strong{mktIndexCRSP:} type `num`. CRSP value weighted market return
}
}
\source{
Unknown
}
\usage{
data(stocksCRSPdaily)
}
\description{
CRSP daily stocks data for 294 stocks for 1993 to 2015
}
\examples{
data(stocksCRSPdaily)

}
\keyword{datasets}
