% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocksCRSPxts.R
\name{stocksCRSPxts}
\alias{stocksCRSPxts}
\title{stocksCRSPxts}
\usage{
stocksCRSPxts(
  data,
  dateRange = c("1993-01-31", "2015-12-31"),
  tickerSet = NULL
)
}
\arguments{
\item{data}{One of data.table objects stocksCRSP, stocksCRSPweekly and
stocksCRSPdaily}

\item{dateRange}{Character vector with two components a start date
and an end date}

\item{tickerSet}{A subset of tickers of the stocks in stocksCRSP}
}
\value{
An xts time series object
}
\description{
Function to extract a set of stocksCRSP data specified by a date
range, and a set of tickers, with convenient defaults
}
\details{
The start date and end date of the dateRange paramter must have
the format "yyyy-mm-dd", and the default is the entire stocksCRSP data
range "1993-01-31" to "2015-12-32". If the tickerSet is NULL, then the 
resulting xts object contains all the stocks in stocksCRSP.
}
\examples{
args(stocksCRSPxts)
tickers4 <- c("DHR","CSL","AVP","AMWD")
dateRange <- c("2011-01-31","2015-12-31")
returns4 <- stocksCRSPxts(stocksCRSP, dateRange = dateRange,
                           tickerSet = tickers4) 
class(returns4)
dim(returns4)
names(returns4)
range(index(returns4))                         
}
