% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strategies}
\alias{strategies}
\title{Hedge Fund Strategies Returns}
\format{
A multivariate xts object
}
\source{
Unknown
}
\usage{
data(strategies)
}
\description{
Monthly returns of 9 hedge fund strategies from 1994 to 2004
}
\examples{
library(PCRA)
data(strategies)
names(strategies)
dim(strategies)
range(index(strategies))
}
\keyword{datasets}
