% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_monthly.R
\name{to_monthly}
\alias{to_monthly}
\title{Function to convert from daily to weekly returns.}
\usage{
to_monthly(daily, index_last = TRUE)
}
\arguments{
\item{daily}{An xts object of daily returns.}

\item{index_last}{Controls whether the return date label will fall on the 
month end date provided in the dataset (usually the final trading date of the
month in most financial data sets) or the calendar month end date 
(potentially a non-trading date). Regardless, the returns are identical 
under two scenarios, but the date may differ for the "month end".}
}
\value{
monthly
}
\description{
to_monthly will convert daily returns to monthly returns.
}
\details{
tbc
}
