% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferCoef.R
\name{transferCoef}
\alias{transferCoef}
\title{Transfer Coefficent}
\usage{
transferCoef(returns, wtVec)
}
\arguments{
\item{returns}{An xts multivariate returns object that contains
the returns of the risk-free T-Bill in the last column}

\item{wtVec}{The weight vector of a constrained MV portfolio}
}
\value{
Numeric value of the TC
}
\description{
Computes the transfer coefficient (TF), which measures
the reduction in mean excess return of an MV portfolio with weights
constraint relative to the mean excess return of an unconstrained
MV portfolio.  This description also holds with portfolios mean
returns replaced by Sharpe ratios.
}
\examples{
args(transferCoef)
}
