% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnOver.R
\name{turnOver}
\alias{turnOver}
\title{Portfolio Turnover}
\usage{
turnOver(weights)
}
\arguments{
\item{weights}{A multivariate xts object of portfolio weights}
}
\value{
A zoo time series object containing T-1 turnover values
}
\description{
Calculates T-1 turn-over values for a times of portfolio
weight vectors from time t = 1 to time t = T, where the turnover from time
t-1 to time t is the sum of the absolute differences between the portfolio
weights at time t-1 and time t.
}
\examples{
args(turnOver)

}
