% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnoverOpt.R
\name{turnoverOpt}
\alias{turnoverOpt}
\title{turnoverOpt}
\usage{
turnoverOpt(
  returns,
  mu.target = NULL,
  wts.initial = NULL,
  toc = NULL,
  long.only = TRUE,
  printD = F,
  printA = F,
  printout = T
)
}
\arguments{
\item{returns}{}

\item{mu.target}{}

\item{wts.initial}{}

\item{toc}{}

\item{long.only}{}

\item{printD}{}

\item{printA}{}

\item{printout}{}
}
\value{
Optimal turnover constrained minVar portfolio, weights, buy weights,
sell weights, turnover, portfolio mean, and portfolio variacne)
}
\description{
turnoverOpt
}
\examples{
args(turnoverOpt)
}
