% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDevPkg.R
\name{update_dev_pkg}
\alias{update_dev_pkg}
\title{Update to Developer version on Github that have access to additional functions and data}
\usage{
update_dev_pkg(
  pkg = "PCRA",
  repo = "https://github.com/robustport/PCRA",
  field = "Version",
  lib = NULL
)
}
\arguments{
\item{pkg}{Default to "PCRA" package name.}

\item{repo}{Default to "https://github.com/robustport/PCRA"}

\item{field}{Default to "Version" field under Package Description file.}

\item{lib}{library path where the package would be installed.}
}
\description{
Update to Developer version on Github that have access to additional functions and data
}
