% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WeeklyReturns1980}
\alias{WeeklyReturns1980}
\title{Weekly Returns 1980}
\format{
Multivariate time series zoo object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business.
}
\usage{
data(WeeklyReturns1980)
}
\description{
Weekly returns of 10 stocks and 90 day T-Bill from November
1980 through Octobber 1981.
}
\examples{
library(PCRA)
data(WeeklyReturns1980)
class(WeeklyReturns1980)
names(WeeklyReturns1980)
head(WeeklyReturns1980,2)
range(index(retDD))
}
\keyword{datasets}
