% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize Data}
\usage{
winsorize(x, fraction = 0.1)
}
\arguments{
\item{x}{A numeric data set}

\item{fraction}{A fraction greater than 0 and less than 0.5}
}
\value{
The Winsorized numeric data
}
\description{
This function Winsorizes a fraction gamma of a numeric data
set.
}
\details{
The Winsorized data is obtained by by setting the gamma smallest
data values equal to the next smallest value, and setting the gamma largest
data values equal to the next largest data value.
}
\examples{
x <- rt(10,8)
winsorize(x,0.2)
}
