% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtsOpt.R
\name{wtsOpt}
\alias{wtsOpt}
\title{wtsOpt}
\usage{
wtsOpt(r, c)
}
\arguments{
\item{r}{}

\item{c}{}
}
\value{
Value of old polynomial approximation
}
\description{
Old polynomial approximation to opt function, to be discarded
}
\examples{
args(wtsOpt)
}
