% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotWts.R
\name{barplotWts}
\alias{barplotWts}
\title{barplotWts}
\usage{
barplotWts(
  wts.efront,
  legend.text = NULL,
  col = NULL,
  ylab = NULL,
  xlab = c("MU", "VOL"),
  bar.ylim = NULL,
  ...
)
}
\arguments{
\item{wts.efront}{Weights along the efficient frontiers}

\item{legend.text}{}

\item{col}{}

\item{ylab}{}

\item{xlab}{}

\item{bar.ylim}{}
}
\description{
Brief description will be added here
}
\examples{
args(barplotWts)
}
