% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotWtsDoug.R
\name{barplotWtsDoug}
\alias{barplotWtsDoug}
\title{Barplot of Efficient Frontier Weights}
\usage{
barplotWtsDoug(
  wts.efront,
  legend.text = NULL,
  col = NULL,
  ylab = NULL,
  xlab = "VOL",
  bar.ylim = NULL,
  ...
)
}
\arguments{
\item{wts.efront}{}

\item{legend.text}{}

\item{col}{}

\item{ylab}{}

\item{xlab}{}

\item{bar.ylim}{}

\item{...}{}
}
\value{
A figure
}
\description{
Barplot of Efficient Frontier Weights
}
\examples{
args(barplotWtsDoug)
}
