% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crsp.returns8}
\alias{crsp.returns8}
\title{CRSP Returns for 8 stocks in 4 cap groups}
\format{
A multivariate xts object
}
\source{
Center for Research in Security Prices, LLC (CRSP), an Affiliate of
the University of Chicago Booth School of Business. NOTE: CRSP data is not
covered by the GPL. Redistribution of the data in any form is not permitted,
and use of the data in derivative works is not permitted without the written
permission of CRSP.
}
\usage{
data(crsp.returns8)
}
\description{
Monthly returns of 8 stocks with tickers GHI, PBCI, MODI, MGJ, MAT,
EMN, AMAT, AMGN, two in each of for capitalization groups from 1997 to 2001
}
\examples{
library(PCRA)
library(zoo)
data(crsp.returns8)
names(crsp.returns8)
dim(crsp.returns8)
range(index(crsp.returns8))
}
\keyword{datasets}
