% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumActMgr.R
\name{cusumActMgr}
\alias{cusumActMgr}
\title{cusumActMgr Statistical Process Control to monitor benchmark-relative portfolio performance}
\usage{
cusumActMgr(
  portfolioName,
  benchmarkName,
  return_df,
  upper_IR = 0.5,
  lower_IR = 0,
  lambda_in = 0.9,
  lambda_out = 0.8,
  winsorize = 4,
  filterStd = FALSE
)
}
\arguments{
\item{portfolioName}{A character string with the name of the portfolio. Required, no default}

\item{benchmarkName}{A character string with the name of the benchmark. Required, no default}

\item{return_df}{An xts object containing the columns \code{portfolioName} and \code{benchmarkName}
of monthly returns. Required, no default}

\item{upper_IR}{Numeric value representing the LIR of a good portfolio, default = 0.5}

\item{lower_IR}{Numeric value representing the LIR of a bad portfolio,  default = 0}

\item{lambda_in}{Exponential weighting constant used when the data seems consistent with
the current estimate of volatility. Default = 0.9}

\item{lambda_out}{Exponential weighting constant used when the data seems inconsistent with
the current estimate of volatility. Default = 0.8}

\item{winsorize}{Numeric value >1, of the multiple of the standard deviation 
at which we winsorize returns to obtain a robust estimate of
the volatility of logarithmic active returns. The default is 4}

\item{filterStd}{Logical value, determines if the estimated volatility of 
logarithmic active returns is filtered with a simple digital 
filter to induce some additional persistence in the estimate.
The default is set to \code{FALSE}}
}
\value{
\code{cusumActMgr} returns a \code{list} containing the following xts objects:
\item{Log_Active_Returns}{Logarithmic active returns of the fund relative to its benchmark}
\item{Annual_Moving_Average}{Vector of annualized moving average annualized returns}
\item{Tracking_Error}{Monthly tracking error of the logarithmic active returns}
\item{Information_Ratios}{Vector of monthly information ratios}
\item{Lindleys_recursion}{Vector of values from Lindley's recursion with a reset after the detection threshold (6.66) is crossed}
\item{Annualized_Cusum_Log_IR}{Vector of the CUSUM of the annualized log information ratios}
\item{Means}{xts matrix of estimated current mean of the fund's returns in the first column,
the benchmark in the second column, and the logarithmic active returns in the third column}
\item{Standard_Deviations}{The xts matrix of estimated standard deviations of
the fund in the first column, the benchmark in the second column, and the
logarithmic active returns in the third column.
The estimated standard deviation is filtered with a simple digital filter to
induce some additional persistence if \code{filterStd = TRUE} is specified}
\item{Protractor_IR}{The xts matrix of the rays of a protractor from
IR = -IR_min in the first column to IR = +IR_max in the seventh column}
\item{Protractor_Log_AR}{The xts matrix of the rays of a protractor from
r = -r_min in the first column to r = +r_max in the seventh column}
\item{Excess_Volatility}{The difference between the annualized standard deviation of the
portfolio and its benchmark}
\item{Annualized_Cusum_Log_AR}{Vector of the CUSUM of the annualized log active returns}
\item{Log_AR_Slopes}{Numeric vector of the seven protractor ray slopes for the log active return plot}
\item{IR_Slopes}{Numeric vector of the seven protractor ray slopes for the information ratio plot}
\item{portfolioName}{The portfolio name passed in by the user}
\item{benchmarkName}{The benchmark name passed in by the user}
\item{returns}{The \code{xts} object of returns passed in by the user}
}
\description{
Monitor the Logarithmic Information Ratio (LIR) of a portfolio relative to its benchmark
and raise an alarm when sufficient evidence accrues to indicate that its current LIR is <0
using an optimal changepoint detection scheme (the CUSUM algorithm).
A named \code{list} of \code{xts} objects and summary vectors is returned.
}
\details{
Assessing the performance of active managers is hard because active returns are noisy.
In addition, their risk must be taken into account, and this is commonly measured by the
standard deviation of active returns. Empirical studies of active managers across a wide range
of asset classes suggest that an Annualized Logarithmic Information Ratio
(LIR = Active Log Return / Std. Dev.(Active Log Return)) of 0.5 over a period of 5 years or more
is exceptional, as markets are efficient. The majority of active managers deliver active
returns and IR close to 0, and even those with a positive IR are at constant risk of having their
added value dissipate. Investors, therefore, must continually estimate the current performance of
their active portfolios and determine when sufficient evidence has accrued to suggest that their
active return and IR have fallen to 0. Put differently, investors need to rapidly detect changes,
particularly negative changes, in the performance of their portfolios

There is a rich literature on changepoint detection, and of the many available algorithms,
the CUSUM (an acronym for CUmulative SUM) stands out, on account of its simplicity,
its robustness to the actual distribution of returns, and the optimal trade-off
between detection time and the rate of false alarms that it offers.
It is closely related to Wald's Sequential Probability Ratio Test (SPRT) but is much simpler
to implement, and requires minimal inputs from the user. In this application, it seeks to
determine when the IR of a portfolio has changed from a good level (default = 0.5 )
to a bad level (default = 0). An alarm is raised when the CUSUM scheme crosses a threshold,
which is chosen to make the average time between false alarms 84 months (7 years).
By way of comparison, the time taken to detect a transition from an LIR of 0.5 to 0 is
41 months (or 3 1/2 years). This is much faster than the traditional t-test, which would take
16 years to obtain a t-statistic of 2. The threshold can be changed to meet a user's needs
}
\examples{
\dontrun{
# Data Preprocessing
df <- read.csv("Example1.csv", header = TRUE, sep = ",", stringsAsFactors = FALSE)
df$Date <- as.yearmon(df$Date, format = "\%m/\%d/\%Y")
df <- as.xts(df[, -1], order.by = df$Date)

# Run the cusumActMgr routine to get a results object
results <- cusumActMgr(portfolioName = colnames(df)[1],
                       benchmarkName = colnames(df)[2], return_df = df)

# Print page 1 of the results
chartCusum(results, select_option = 1,
           print_to_screen = TRUE,
           print_to_png = TRUE,  png_fn = "Mgr vs Bmk Page 1.png",
           print_to_pdf = FALSE, pdf_fn = "Mgr vs Bmk Page 1.pdf")

# Print page 2 of the results
chartCusum(results, select_option = 2,
           print_to_screen = TRUE,
           print_to_png = TRUE,  png_fn = "Mgr vs Bmk Page 2.png",
           print_to_pdf = FALSE, pdf_fn = "Mgr vs Bmk Page 2.pdf")
}
}
\references{
Philips, T. K., Yashchin, E. and Stein, D. M. (2003). Using Statistical Process Control
to Monitor Active Managers, Journal of Portfolio Management, Fall 2003, pp. 86-94
}
\author{
Chindhanai Uthaisaad (GSoC 2017) Vinav Singh Sancheti (2021) and Thomas Philips (2025)
}
