% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divHHI.R
\name{divHHI}
\alias{divHHI}
\title{HHI Based Diversification Index}
\usage{
divHHI(weights)
}
\arguments{
\item{weights}{A numeric vector of portfolio weights}
}
\value{
Time series of portfolio diversification values
}
\description{
divHHI calculates a portfolio diversification index DIV. The DIV
is equal to 1 minus the Herfindahl-Hirschman Index (HHI), which is defined
as the sum of the squared portfolio weights. The maximum HHI of a long-only
portfolio is 1, which occurs when all of the portfolio's investment is in a
single asset, and correspondingly HHI = 0.
}
\examples{
args(divHHI)
}
