% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huberRhoPsiWtsEff.R
\name{effFromConstHuber}
\alias{effFromConstHuber}
\title{Efficiency From Constant Huber Psi}
\usage{
effFromConstHuber(cc = 1.345)
}
\arguments{
\item{cc}{a numeric constant}

\item{x}{a numeric vector}
}
\description{
Efficiency From Constant Huber Psi
}
\examples{
args(effFromConstHuber)
}
\seealso{
\code{\link{}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{huber}
