% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRAData.R
\name{getPCRAData}
\alias{getPCRAData}
\title{Download CRSP and SPGMI Data}
\usage{
getPCRAData(dataset = "stocksCRSPweekly", cache = TRUE, refresh = FALSE)
}
\arguments{
\item{dataset}{a valid dataset name (see details)}

\item{cache}{whether to cache the operation so that when calling the function
using the same dataset it will be loaded from cache rather than re-downloading}

\item{refresh}{whether to re-download a cached dataset}
}
\value{
An object of class \dQuote{data.table}.
}
\description{
Downloads stocksCRSPweekly and stocksCRSPdaily
}
\details{
The following are valid names of datasets available:
\itemize{
\item{"stocksCRSP"}{"Details are available in the stocksCRSP man page"}
\item{"stocksCRSPdaily"}{"Details are the same as for stocksCRSP except
the frequency is now daily"}
\item{"stocksCRSPweekly"}{""Details are the same as for stocksCRSP except
the frequency is now weekly""}
}
}
