% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRAData.R
\name{getPCRAData}
\alias{getPCRAData}
\title{PCRA Sample Datasets}
\usage{
getPCRAData(dataset = "crsp.returns8", cache = TRUE, refresh = FALSE)
}
\arguments{
\item{dataset}{a valid dataset name (see details)}

\item{cache}{whether to cache the operation so that when calling the function
using the same dataset it will be loaded from cache rather than re-downloading}

\item{refresh}{whether to re-download a cached dataset}
}
\value{
An object of class \dQuote{data.table}.
}
\description{
Downloads PCRA sample datasets
}
\details{
The following are valid names of datasets available:
\itemize{
\item{"stocksCRSP"}{"Add Description"}
\item{"stocksCRSPdaily"}{"Add Description"}
\item{"stocksCRSPweekly"}{"Add Description"}
\item{"crsp.returns8"}{"Add Description"}
\item{"crsp.short"}{"Add Description"}
\item{"factorsSPGMI"}{"Add Description"}
\item{"FRBrates1934to2014"}{"Add Description"}
\item{"smallcapM"}{"Add Description"}
\item{"smallcapW"}{"Add Description"}
\item{"largecapM"}{"Add Description"}
\item{"largecapW"}{"Add Description"}
\item{"normalVsHectic"}{"Add Description"}
\item{"ret4withOutliers"}{"Add Description"}
\item{"strategies"}{"Add Description"}
\item{"WeeklyReturns1980"}{"Add Description"}
\item{"wtsGmvLS"}{"Add Description"}
\item{"gfunds5"}{"Add Description"}
\item{"invensysEPS"}{"Add Description"}
}
}
