% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathEfront.R
\name{mathEfront}
\alias{mathEfront}
\title{mathEfront}
\usage{
mathEfront(
  returns,
  mu.max = NULL,
  sigma.max = NULL,
  rf = 0.005,
  rf.line = T,
  stocks = T,
  stock.names = T,
  values = T,
  npoints = 100,
  digits = NULL
)
}
\arguments{
\item{returns}{}

\item{mu.max}{}

\item{sigma.max}{}

\item{rf}{}

\item{rf.line}{}

\item{stocks}{}

\item{stock.names}{}

\item{values}{}

\item{npoints}{}

\item{digits}{}
}
\value{
Efficient frontier values using returns to compute mean vector
and covariance matrix
}
\description{
mathEfront
}
\examples{
args(mathEfront)
}
