% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathEfrontRiskyMuCov.R
\name{mathEfrontRiskyMuCov}
\alias{mathEfrontRiskyMuCov}
\title{Efficient Frontier}
\usage{
mathEfrontRiskyMuCov(
  muRet,
  volRet,
  corrRet,
  npoints = 100,
  display = T,
  efront.only = T,
  print = F,
  digits = NULL
)
}
\arguments{
\item{muRet}{}

\item{volRet}{}

\item{corrRet}{}

\item{npoints}{}

\item{display}{}

\item{efront.only}{}

\item{print}{}

\item{digits}{}
}
\description{
Efficient Frontier based on user specified mean vector and
covariance matrix
}
\examples{
args(mathEfrontRiskyMuCov)
}
