% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathEfrontRiskyMuCov.R
\name{mathEfrontRiskyMuCov}
\alias{mathEfrontRiskyMuCov}
\title{Efficient Frontier}
\usage{
mathEfrontRiskyMuCov(
  muRet,
  volRet,
  corrRet,
  npoints = 100,
  display = T,
  efront.only = T,
  print = F,
  digits = NULL
)
}
\arguments{
\item{muRet}{Numeric vector as asset returns}

\item{volRet}{Numeric vector of asset volatilities}

\item{corrRet}{Correlation matrix of asset returns}

\item{npoints}{Number of points on efficient frontier, default 100}

\item{display}{Logical variable, default TRUE}

\item{efront.only}{Logical variable, default TRUE}

\item{print}{Logical variable, default FALSE}

\item{digits}{}
}
\value{
Plot of efficient frontier
}
\description{
Computes a frontier or efficient frontier based on user
specified mean vector and covariance matrix.  Default is to compute the
efficient frontier and plot it.  Optionally the mean and volatility values
of the frontier or efficient frontier is returned at user specified number
of significant digits.
}
\examples{
args(mathEfrontRiskyMuCov)
}
