% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathWtsEfrontRisky.R
\name{mathWtsEfrontRisky}
\alias{mathWtsEfrontRisky}
\title{Portfolio Weights on an Efficient Frontier}
\usage{
mathWtsEfrontRisky(returns, mu.efront, digits = NULL)
}
\arguments{
\item{returns}{A multivariate xts object of arithmetic asset returns}

\item{mu.efront}{A vector of specified efficient frontier mean returns}

\item{digits}{If null no rounding of weights, else number of significant
digits for rounding}
}
\value{
A matrix whose first row contains the volatilities along the efficient
at specified mean returns contained in the second row, and the remaining
n rows contain contain the n weight vectors at each of the efficient
frontier mean return values
}
\description{
Computes weights vector along an efficient frontier, obtained from a time
series of portfolio returns at specified portfolio mean return values
}
\examples{
args(mathWtsEfrontRisky)
}
