% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanReturns4Types.R
\name{meanReturns4Types}
\alias{meanReturns4Types}
\title{Four Types of Mean Returns}
\usage{
meanReturns4Types(return, robust = FALSE, eff = 0.95)
}
\arguments{
\item{return}{An xts object or a numeric vector of returns}

\item{robust}{A logical value controlling whether a classical or robust
sample mean and standard deviation is computed. Default is FALSE}

\item{eff}{Normal distribution efficiency of RobStatTM function locScaleM()
used for computing a robust location estimate}
}
\value{
fourMeans numeric values of the four means in the Description
}
\description{
Computation of arithmetic mean, logarithmic mean, geometric
mean, and an approximate geometric mean.
}
\examples{
args(meanReturns4Types)
}
