% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanReturns4Types.R, R/meansReturns4Report.R
\name{meansReturns4Report}
\alias{meansReturns4Report}
\title{Four types of returns means for reporting}
\usage{
meansReturns4Report(ret, robust = FALSE, eff = 0.95)

meansReturns4Report(ret, robust = FALSE, eff = 0.95)
}
\arguments{
\item{robust}{A logical value controlling whether a classical or robust sample
mean and standard deviation is computed. Default is FALSE}

\item{x}{An xts or numeric vector of return}
}
\value{
fourMeans

fourMeans
}
\description{
Computes arithmetic mean, logarithmic mean, geometric mean and an approximate
geometric mean.

Computes arithmetic mean, logarithmic mean, geometric mean and an approximate
geometric mean.
}
\examples{
args(meansReturns4Report)
args(meansReturns4Report)
}
