% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minVarCashRisky.R
\name{minVarCashRisky}
\alias{minVarCashRisky}
\title{A Simple Minimum Variance Portfolio}
\usage{
minVarCashRisky(returns, mu0, min = -1000, max = +1000)
}
\arguments{
\item{returns}{an xts multivariate returns object}

\item{mu0}{a specified minimum variance portfolio mean return}

\item{min}{a minimum weight constrain}

\item{max}{a maximum weight constrain}
}
\value{
A list of the minimum variance portfolio numeric weights, 
mean value, volatility, and Sharpe Ratio.
}
\description{
Given a time series of returns that includes a 
risk-free rate, a target mean return and set of weights constraints,
this function the weights of a minimum variance portfolio
}
\examples{
args(minVarCashRisky)
}
