% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRA-package.R
\docType{package}
\name{PCRA-package}
\alias{PCRA}
\alias{PCRA-package}
\title{PCRA: Companion to Portfolio Construction and Risk Analysis}
\description{
A collection of functions and data sets to support teaching a quantitative finance MS level course on Robust Portfolio Construction and Risk Analysis. The package is unique in providing functions for both classical and modern analytic methods. The latter includes robust regression and robust covariance matrix estimation methods that are not much influenced by outliers, minimum downside risk portfolio construction, and computation of accurate standard errors for risk and performance estimators with non-normal and serially dependent returns. The data sets include cross-sections of stock data from the Center for Research on Security Prices, LLC (CRSP), and factor exposures for those stocks from S&P Global.
}
\author{
\strong{Maintainer}: Doug Martin \email{martinrd3d@gmail.com}

Authors:
\itemize{
  \item Kirk Li \email{cocokecoli@gmail.com} [contributor]
}

Other contributors:
\itemize{
  \item Alexios Galanos \email{alexios@4dscape.com} [contributor]
  \item Jon Spinney \email{Jon.Spinney@vestcor.org} [contributor]
}

}
\keyword{internal}
