% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRA-package.R
\docType{package}
\name{PCRA-package}
\alias{PCRA}
\alias{PCRA-package}
\title{PCRA: Companion to Portfolio Construction and Risk Analysis}
\description{
PCRA is a companion package to the text Portfolio Construction and Risk Analysis.
}
\keyword{internal}
