% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRA-package.R
\docType{package}
\name{PCRA-package}
\alias{PCRA}
\alias{PCRA-package}
\title{PCRA: Companion to Portfolio Construction and Risk Analysis}
\description{
PCRA is a companion package to the text Robust Portfolio Construction and Risk Analysis.
}
\author{
\strong{Maintainer}: Doug Martin \email{martinrd3d@gmail.com}

Other contributors:
\itemize{
  \item Jon Spinney \email{Jon.Spinney@vestcor.org} [contributor]
  \item Kirk Li \email{cocokecoli@gmail.com} [contributor]
}

}
\keyword{internal}
