% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRA-package.R
\docType{package}
\name{PCRA-package}
\alias{PCRA}
\alias{PCRA-package}
\title{PCRA: Companion to Portfolio Construction and Risk Analysis}
\description{
PCRA is a companion package to the text Robust Portfolio Construction and Risk Analysis authored by R. Douglas Martin, Thomas K. Philips, Bernd Scherer, Stoyan Stoyanov and Kirk Li, to be published by Springer in 2023. The package includes cross-sections of data for approximately 300 stocks from the Center for Research on Security Prices, LLC (CRSP), and 14 corresponding factor exposures from S&P Global. The book covers classical and modern robust methods of portfolio construction and risk analysis with a reproducibality focus. The data and function in the package support this focus. Extensive R scripts based on package functions for each chapter reproduce most of the tables and figures in the chapter.
}
\author{
\strong{Maintainer}: Doug Martin \email{martinrd3d@gmail.com}

Authors:
\itemize{
  \item Kirk Li \email{cocokecoli@gmail.com} [contributor]
}

Other contributors:
\itemize{
  \item Alexios Galanos \email{alexios@4dscape.com} [contributor]
  \item Jon Spinney \email{Jon.Spinney@vestcor.org} [contributor]
}

}
\keyword{internal}
