% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRA-package.R
\docType{package}
\name{PCRA-package}
\alias{PCRA}
\alias{PCRA-package}
\title{PCRA: Companion to Portfolio Construction and Risk Analysis}
\description{
A collection of functions and data sets that support teaching a quantitative finance MS level course on Portfolio Construction and Risk Analysis, and the writing of a textbook for such a course. The package is unique in providing several real-world data sets that may be used for problem assignments and student projects. The data sets include cross-sections of stock data from the Center for Research on Security Prices, LLC (CRSP), corresponding factor exposures data from S&P Global, and several SP500 data sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robustport/PCRA}
}

}
\author{
\strong{Maintainer}: Doug Martin \email{martinrd3d@gmail.com}

Authors:
\itemize{
  \item Kirk Li \email{cocokecoli@gmail.com} [contributor]
}

Other contributors:
\itemize{
  \item Alexios Galanos \email{alexios@4dscape.com} [contributor]
  \item Jon Spinney \email{Jon.Spinney@vestcor.org} [contributor]
  \item Thomas Philips \email{tkpmep@gmail.com} [contributor]
}

}
\author{
\strong{Maintainer}: Doug Martin \email{martinrd3d@gmail.com}

Other contributors:
\itemize{
  \item Jon Spinney \email{Jon.Spinney@vestcor.org} [contributor]
  \item Kirk Li \email{cocokecoli@gmail.com} [contributor]
}

}
\keyword{internal}
