% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLSandRobustSFM.R
\name{plotLSandRobustSFM}
\alias{plotLSandRobustSFM}
\title{LS and Robust mOpt Single Factor Model (SFM) Fits and Plot}
\usage{
plotLSandRobustSFM(
  x,
  family = "mopt",
  efficiency = 0.95,
  mainText = NULL,
  ylimits = NULL,
  legendPos = "topleft",
  goodOutlier = F,
  makePct = FALSE
)
}
\arguments{
\item{x}{A univariate xts object.}

\item{family}{Robust loss function choice.}

\item{efficiency}{Normal distribution estimator efficiency.}

\item{mainText}{Main title, if any.}

\item{ylimits}{Vertical axis limits.}

\item{legendPos}{Legend position.}

\item{goodOutlier}{}

\item{makePct}{}
}
\value{
The plot with straight line fits and legend containing coefficient
         estimates and their standard errors
}
\description{
LS and Robust mOpt Single Factor Model (SFM) Fits and Plot
}
