% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnsCRSPxts.R
\name{returnsCRSPxts}
\alias{returnsCRSPxts}
\title{returnsCRSPxts}
\usage{
returnsCRSPxts(stocksData)
}
\arguments{
\item{stocksData}{The data.table created by selectCRSPandSPGMI}
}
\value{
A multivariate times series object described above
}
\description{
This function uses the data.table created by selectCRSPandSPGMI to create
a multivarite xts time series object that contains the returns of the stocks
with tickers in the TickerLast component of the data.table, the returns of
the MktIndexCRSP, and the returns of the Ret13WkBill.
}
\examples{
library(PCRA)
library(data.table)
stockItems <- c("Date","TickerLast","CapGroupLast","Return","MktIndexCRSP",
               "Ret13WkBill")
dateRange <- c("1997-01-31","2002-12-31")
stocksDT <- selectCRSPandSPGMI("monthly",dateSet = dateRange, stockItems =
                                 stockItems, factorItems = NULL)
stocksData <- stocksDT[[1]][CapGroupLast == "SmallCap"]
ret <- returnsCRSPxts(stocksData)
tickers <- unique(stocksData[,TickerLast])
tickers10 <- tickers[11:20]
colnames <- c(tickers10,"Market","RiskFree")
head(ret[,colnames],1)
}
