% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocksCRSPxts.R
\name{stocksCRSPxts}
\alias{stocksCRSPxts}
\title{stocksCRSPxts}
\usage{
stocksCRSPxts(
  data,
  dateRange = c("1993-01-31", "2015-12-31"),
  tickerSet = NULL
)
}
\arguments{
\item{data}{One of data.table objects stocksCRSP, stocksCRSPweekly and
stocksCRSPdaily.  Referred to collectively as the stocksCRSP data.}

\item{dateRange}{Character vector with two components a start date
and an end date using format "yyyy-mm-dd".  Default is the entire
stocksCRSP data dates range c("1993-01-31","2015-12-32")}

\item{tickerSet}{A subset of tickers of the stocks in stocksCRSP. The
default is tickerSet = NULL, which results in selection of all stocks
in stocksCRSP.}
}
\value{
An xts time series object
}
\description{
Function to extract a set of stocksCRSP data specified by a date
range, and a set of tickers, with convenient defaults, and convert it
to an xts time series object
}
\examples{
library(PCRA)
library(data.table)
args(stocksCRSPxts)
tickers4 <- c("DHR","CSL","AVP","AMWD")
dateRange <- c("2011-01-31","2015-12-31")
returns4 <- stocksCRSPxts(stocksCRSP, dateRange = dateRange,
                           tickerSet = tickers4) 
class(returns4)
dim(returns4)
names(returns4)
range(index(returns4))                         
}
