% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnOver.R
\name{turnOver}
\alias{turnOver}
\title{Calculate Portfolio Turnover}
\usage{
turnOver(weights)
}
\arguments{
\item{weights}{multivariate xts object of portfolio weights}
}
\value{
Time series of turnover values
}
\description{
turnOver is the function to calculate a times series of n-1 turn-over
values for a times of n portfolio weight vectors, where the turnover
from time i to time i+1 is the sum of the absolute differences 
between the portfolio weights at time i and time i+1
}
\examples{
args(turnOver)
}
