% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnOver.R
\name{turnOver}
\alias{turnOver}
\title{Portfolio Turnover}
\usage{
turnOver(weights)
}
\arguments{
\item{weights}{Multivariate xts object of portfolio weights}
}
\value{
Time series of T-1 turnover values
}
\description{
Calculates n-1 turn-over values for a times of n portfolio
weight vectors, where the turnover from time t to time t+1 is the sum of the
absolute differences between the portfolio weights at time t and time t+1.
}
\examples{
args(turnOver)

}
