% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorMean.R
\name{winsorMean}
\alias{winsorMean}
\title{Winsorized Mean}
\usage{
winsorMean(x, winFrac = 0, na.rm = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{winFrac}{Fraction of data to be Winsorized}

\item{na.rm}{Logical variable with default FALSE}

\item{...}{Pass-through parameters}
}
\value{
Numeric value of Winsorized mean
}
\description{
Winsorized Mean
}
\examples{
args(winsorMean)
}
