% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorMean.R
\name{winsorMean}
\alias{winsorMean}
\title{Winsorized Mean}
\usage{
winsorMean(x, trim = 0, na.rm = FALSE, ...)
}
\arguments{
\item{x}{numeric vector}

\item{trim}{fraction of data to be Winsorized}

\item{na.rm}{}

\item{...}{}
}
\description{
Winsorized Mean
}
\examples{
args(winsorMean)
}
